Fix Pylint warnings, update docs and move utils_test to tests
