+++ docs/modules/utils_test.rst
@@ -1,12 +1,16 @@
-:mod:`art.utils_test`
-=====================
-.. automodule:: art.utils_test
+:mod:`tests.utils_test`
+=======================
+.. automodule:: tests.utils_test
 
 Trained Models for Unittests, MNIST
 -----------------------------------
 .. autofunction:: get_classifier_tf
+.. autofunction:: get_classifier_tf_v2
 .. autofunction:: get_classifier_kr
+.. autofunction:: get_classifier_kr_tf
+.. autofunction:: get_classifier_kr_tf_binary
 .. autofunction:: get_classifier_pt
+.. autofunction:: get_classifier_bb
 
 Trained Models for Unittests, Iris
 ----------------------------------
