+++ tests/attacks/test_adversarial_patch.py
@@ -25,7 +25,7 @@ from sklearn.tree import DecisionTreeClassifier
 
 from art.attacks import AdversarialPatch
 from art.utils import load_dataset, master_seed
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt, get_iris_classifier_kr
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt, get_iris_classifier_kr
 from art.classifiers.scikitlearn import ScikitlearnDecisionTreeClassifier
 
 logger = logging.getLogger(__name__)
