+++ tests/attacks/test_carlini.py
@@ -27,8 +27,8 @@ import numpy as np
 from art.attacks import CarliniL2Method, CarliniLInfMethod
 from art.classifiers import KerasClassifier
 from art.utils import load_dataset, random_targets, master_seed, to_categorical
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from tests.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger(__name__)
 
@@ -304,9 +304,8 @@ class TestCarlini(unittest.TestCase):
 
     def test_scikitlearn_L2(self):
         from sklearn.linear_model import LogisticRegression
-        from sklearn.svm import SVC, LinearSVC
 
-        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression, ScikitlearnSVC
+        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression
 
         scikitlearn_test_cases = {LogisticRegression: ScikitlearnLogisticRegression}  # ,
         # SVC: ScikitlearnSVC,
@@ -587,9 +586,8 @@ class TestCarlini(unittest.TestCase):
 
     def test_scikitlearn_LInf(self):
         from sklearn.linear_model import LogisticRegression
-        from sklearn.svm import SVC, LinearSVC
 
-        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression, ScikitlearnSVC
+        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression
 
         scikitlearn_test_cases = {LogisticRegression: ScikitlearnLogisticRegression}  # ,
         # SVC: ScikitlearnSVC,
