+++ tests/attacks/test_fast_gradient.py
@@ -26,8 +26,8 @@ from art.attacks import FastGradientMethod
 from art.classifiers import KerasClassifier
 from art.defences import FeatureSqueezing
 from art.utils import load_dataset, get_labels_np_array, master_seed, random_targets
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from tests.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger(__name__)
