+++ tests/attacks/test_hop_skip_jump.py
@@ -26,8 +26,8 @@ import numpy as np
 from art.attacks import HopSkipJump
 from art.classifiers import KerasClassifier
 from art.utils import load_dataset, random_targets, master_seed
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from tests.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger(__name__)
