+++ tests/attacks/test_newtonfool.py
@@ -25,8 +25,8 @@ import numpy as np
 from art.attacks import NewtonFool
 from art.classifiers import KerasClassifier
 from art.utils import load_dataset, master_seed
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from tests.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger(__name__)
 
@@ -200,9 +200,8 @@ class TestNewtonFool(unittest.TestCase):
 
     def test_scikitlearn(self):
         from sklearn.linear_model import LogisticRegression
-        from sklearn.svm import SVC, LinearSVC
 
-        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression, ScikitlearnSVC
+        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression
 
         scikitlearn_test_cases = {LogisticRegression: ScikitlearnLogisticRegression}  # ,
         # SVC: ScikitlearnSVC,
