+++ tests/attacks/test_projected_gradient_descent.py
@@ -25,8 +25,8 @@ import numpy as np
 from art.attacks import ProjectedGradientDescent
 from art.classifiers import KerasClassifier
 from art.utils import load_dataset, get_labels_np_array, master_seed, random_targets
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from tests.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger(__name__)
