+++ tests/attacks/test_saliency_map.py
@@ -25,8 +25,8 @@ import numpy as np
 from art.attacks import SaliencyMapMethod
 from art.classifiers import KerasClassifier
 from art.utils import load_dataset, get_labels_np_array, to_categorical, master_seed
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from tests.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger(__name__)
 
@@ -298,9 +298,8 @@ class TestSaliencyMap(unittest.TestCase):
 
     def test_scikitlearn(self):
         from sklearn.linear_model import LogisticRegression
-        from sklearn.svm import SVC, LinearSVC
 
-        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression, ScikitlearnSVC
+        from art.classifiers.scikitlearn import ScikitlearnLogisticRegression
 
         scikitlearn_test_cases = {LogisticRegression: ScikitlearnLogisticRegression}  # ,
         # SVC: ScikitlearnSVC,
