+++ tests/attacks/test_universal_perturbation.py
@@ -25,8 +25,8 @@ import numpy as np
 from art.attacks import UniversalPerturbation
 from art.classifiers import KerasClassifier
 from art.utils import load_dataset, master_seed
-from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from tests.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from tests.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger(__name__)
