+++ tests/classifiers/test_keras_tf.py
@@ -42,7 +42,7 @@ from art.classifiers import KerasClassifier
 from art.classifiers.keras import generator_fit
 from art.defences import FeatureSqueezing, JpegCompression, SpatialSmoothing
 from art.utils import load_dataset, master_seed
-from art.utils_test import get_classifier_kr_tf
+from tests.utils_test import get_classifier_kr_tf
 from art.data_generators import KerasDataGenerator
 
 logger = logging.getLogger(__name__)
