+++ tests/wrappers/test_wrapper.py
@@ -25,7 +25,7 @@ import numpy as np
 
 from art.wrappers.wrapper import ClassifierWrapper
 from art.utils import load_mnist, master_seed
-from art.utils_test import get_classifier_kr
+from tests.utils_test import get_classifier_kr
 
 logger = logging.getLogger(__name__)
